+++ modules/router/src/services/config.ts
@@ -1,11 +1,13 @@
 import { Result } from "@connext/vector-types";
+import { getAddress } from "@ethersproject/address";
 
 import { config, RebalanceProfile } from "../config";
 import { ConfigServiceError } from "../errors";
 
 export const getRebalanceProfile = (chainId: number, assetId: string): Result<RebalanceProfile, ConfigServiceError> => {
+  const asset = getAddress(assetId);
   const rebalanceProfile = config.rebalanceProfiles.find(
-    (profile) => profile.assetId === assetId && profile.chainId === chainId,
+    (profile) => profile.assetId === asset && profile.chainId === chainId,
   );
   if (!rebalanceProfile) {
     return Result.fail(