+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,15 @@
 
 ## Next Release
 
+## 0.1.1
+
+- \[all\] Improve logging
+- \[contracts\] Add etherscan verification plugin
+- \[router\] Correctly reclaim from the sender's channel
+- \[contracts\] Check if channel is deployed right before deploying
+- \[router\] Checksum config assetIds
+- \[engine\] Checksum user input assetIds
+
 ## 0.1.0
 
 - \[all\] Mainnet release :tada: